requirement.txt
# Core Dependencies
python-dotenv>=1.0.0
groq>=0.4.0
pandas>=2.0.0
streamlit>=1.28.0

# Job Scraping
python-jobspy>=1.1.0

# PDF Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# Utilities
requests>=2.31.0
